pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "markdown>=3.8",
    "python-frontmatter>=1.1.0",
]
